Enable writing after the 2G boundary
This patch fixes a bug in readwrite.c which truncated the 'from' field
in nbd requests.  It was casting them down from an off64_t to an int.

+++ src/readwrite.c
@@ -56,7 +56,7 @@ fail:
 	return 0;
 }
 
-void fill_request(struct nbd_request *request, int type, int from, int len)
+void fill_request(struct nbd_request *request, int type, off64_t from, int len)
 {
 	request->magic  = htobe32(REQUEST_MAGIC);
 	request->type   = htobe32(type);
